+++ api/operator/v1beta1/vmagent_types.go
@@ -403,18 +403,17 @@ func (cr *VMAgent) Validate() error {
 		}
 		if sc.TLSConfig != nil {
 			if err := sc.TLSConfig.Validate(); err != nil {
-				return fmt.Errorf("bad tlsConfig for scrape class %q: %w", sc.Name, err)
+				return fmt.Errorf("incorrect tlsConfig for scrape class %q: %w", sc.Name, err)
 			}
 		}
-		if len(sc.Relabelings) > 0 {
-			if err := checkRelabelConfigs(sc.Relabelings); err != nil {
-				return fmt.Errorf("bad relabelings for scrape class %q: %w", sc.Name, err)
-			}
+		if err := sc.OAuth2.validate(); err != nil {
+			return fmt.Errorf("incorrect oauth2 for scrape class %q: %w", sc.Name, err)
 		}
-		if len(sc.MetricRelabelings) > 0 {
-			if err := checkRelabelConfigs(sc.MetricRelabelings); err != nil {
-				return fmt.Errorf("bad metricRelabelings for scrape class %q: %w", sc.Name, err)
-			}
+		if err := sc.Authorization.validate(); err != nil {
+			return fmt.Errorf("incorrect authorization for scrape class %q:: %w", sc.Name, err)
+		}
+		if err := sc.validate(); err != nil {
+			return fmt.Errorf("incorrect relabeling for scrape class %q:: %w", sc.Name, err)
 		}
 	}
 	return nil
@@ -495,27 +494,8 @@ type ScrapeClass struct {
 	// +optional
 	Default *bool `json:"default,omitempty"`
 
-	// tlsConfig defines the TLS settings to use for the scrape. When the
-	// scrape objects define their own CA, certificate and/or key, they take
-	// precedence over the corresponding scrape class fields.
-	//
-	// For now only the `caFile`, `certFile` and `keyFile` fields are supported.
-	//
-	// +optional
-	TLSConfig *TLSConfig `json:"tlsConfig,omitempty"`
-
-	// authorization section for the ScrapeClass.
-	// It will only apply if the scrape resource doesn't specify any Authorization.
-	// +optional
-	Authorization *Authorization `json:"authorization,omitempty"`
-
-	// Relabelings defines the relabeling rules to apply to all scrape targets.
-	// +optional
-	Relabelings []*RelabelConfig `json:"relabelings,omitempty"`
-
-	// MetricRelabelings defines the relabeling rules to apply to all samples before ingestion.
-	// +optional
-	MetricRelabelings []*RelabelConfig `json:"metricRelabelings,omitempty"`
+	EndpointAuth        `json:",inline"`
+	EndpointRelabelings `json:",inline"`
 
 	// AttachMetadata defines additional metadata to the discovered targets.
 	// When the scrape object defines its own configuration, it takes